[ui] Move Search component into search/components directory
